requirements.txt
openai==1.44.1
slack_sdk==3.32.0
slack_bolt==1.20.1
python-dotenv==1.0.1
boto3==1.35.17
